[FEAT] Move `argilla` to `argilla-v1` and define the new SDK as `argilla` (#4975)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

This is a WIP PR to evaluate using the new argilla SDK as the `argilla`
module in Python. The changes included in this PR do not affect the
current `argilla-sdk` project. I've created a brand new `argilla`
project quite simple. The project `argilla` manages dependencies for
both SDK versions under the same package.

Having this new `argilla` package is not a requirement. We can rename
the current `argilla-sdk` to `argilla` and include the `argilla-v1`
dependency and the `v1` module. We've already discussed having a bundle
package for adding and managing extensions or contribs packages through
argilla:
```bash
pip install argilla[extensions]
pip install argilla[contrib]
```
instead of
```bash
pip install argilla-extensions
pip install argilla-contrib
````

But I think this can be tackled once we have more packages other than
argilla. So, my proposal for the v2 release is to move `argilla-sdk` to
the official `argilla` project.


## Update 

The decision will be to rename `argilla-sdk` to `argilla` and expose the
`v1` as a dependency.. See PR
https://github.com/argilla-io/argilla/pull/4975 to see some of this
work.


Refs: https://github.com/argilla-io/argilla/issues/4889

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] New feature (non-breaking change which adds functionality)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [ ] Test A
- [ ] Test B

**Checklist**

- [ ] I added relevant documentation
- [ ] I followed the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)