[DOCS] add migration notebook to docs format (#5002)
This PR transfer the notebook guide by @frascuchon and @nataliaElv over
to the documentation format, which makes it easier to follow.

You can find the page here:
http://localhost:8000/argilla-python/guides/how_to_guides/migrate_from_legacy_datasets/

n.b. the page is not available via the menu. @davidberenstein1957 we
should figure out navigation presentation for this page.

The guide looks like this:


![image](https://github.com/argilla-io/argilla/assets/19620375/e4e00da8-6117-4f45-9663-1a8c8be9e987)

---------

Co-authored-by: Francisco Aranda <francis@argilla.io>
Co-authored-by: David Berenstein <david.m.berenstein@gmail.com>
Co-authored-by: Ben Burtenshaw <benjamin.burtenshaw@huggingface.co>