docs/document how to setup hf oauth (#5125)
# Pull Request Template
<!-- Please include a summary of the changes and the related issue.
Please also include relevant motivation and context. List any
dependencies that are required for this change. -->

COMMENTS:
- I'm not sure about the index shown due to the installation.md file (as
it encompass the library installation, mention the server and some
developer info), maybe it would be useful to also add a basic docker.md
as they are the two basic ways of deploying.
- I also thought of mentioning that they can use Oauth with other apps,
but while we don't add other type of deployments (community ones) I
decided not to confuse.

Closes #5091 

**Type of change**
<!-- Please delete options that are not relevant. Remember to title the
PR according to the type of change -->

- Documentation update

**How Has This Been Tested**
<!-- Please add some reference about how your feature has been tested.
-->
mkdocs serve

**Checklist**
<!-- Please go over the list and make sure you've taken everything into
account -->

- I added relevant documentation

---------

Co-authored-by: David Berenstein <david.m.berenstein@gmail.com>