[FEATURE] encode class labels as names not ints (#5474)
# Description
<!-- Please include a summary of the changes and the related issue.
Please also include relevant motivation and context. List any
dependencies that are required for this change. -->

This PR add to the settings from_hub PR by supporting ClassLabels coming
from datasets package.

It renders a label question with names based on the ClassLabel.

It uncasts the 'label.suggestion' to a corresponding class label.

**Type of change**
<!-- Please delete options that are not relevant. Remember to title the
PR according to the type of change -->

- New feature (non-breaking change which adds functionality)