[BUGFIX] argilla server: Prevent convert `ChatFieldValue` objects (#5605)
# Description
<!-- Please include a summary of the changes and the related issue.
Please also include relevant motivation and context. List any
dependencies that are required for this change. -->

Changes introduced in https://github.com/argilla-io/argilla/pull/5600
include a but when the fields in the `RecordCreate` schema already
contain `ChatFieldValue` object. This is the cause of [failing
integration tests in the develop
branch](https://github.com/argilla-io/argilla/actions/runs/11380531039/job/31660006108).

The custom validator must skip these objects even if the `pre` flag is
set to True.

This PR fixes the error.

**Type of change**
<!-- Please delete options that are not relevant. Remember to title the
PR according to the type of change -->

- Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**
<!-- Please add some reference about how your feature has been tested.
-->

**Checklist**
<!-- Please go over the list and make sure you've taken everything into
account -->

- I added relevant documentation
- I followed the style guidelines of this project
- I did a self-review of my code
- I made corresponding changes to the documentation
- I confirm My changes generate no new warnings
- I have added tests that prove my fix is effective or that my feature
works
- I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)