feat(helm): add support for default storage class in PVCs
I updated the Helm chart to allow the use of a default storage class
when none is explicitly specified. This change is important because some
Kubernetes environments, like k3d, do not have a 'standard' storage
class by default.
Without this change, the setup could fail in such environments, as
Kubernetes would attempt to use an undefined or missing storage class.
By not explicitly setting a storage class, I allow Kubernetes to default
to whatever storage class is configured for the cluster.

This ensures the chart works across various setups, including those that
don't predefine a 'standard' storage class.